@@ -74,7 +74,6 @@ namespace ReservationAPI.Controllers
                 Street = model.Street,
                 City = model.City,
                 Image = model.Image,
-                //Friends = new List<Friend>(),
                 Reservations = new List<Reservation>(),
                 PassportNo = model.PassportNo
             };
@@ -98,7 +97,7 @@ namespace ReservationAPI.Controllers
             catch (Exception ex)
             {
                 Console.WriteLine($"ERROR in registering new user. -> {ex.Message}");
-                throw ex;
+                throw;
             }
         }
 
@@ -160,7 +159,7 @@ namespace ReservationAPI.Controllers
             catch (Exception ex)
             {
                 Console.WriteLine($"ERROR in registering new guest. -> {ex.Message}");
-                throw ex;
+                return null;
             }
         }
 
@@ -179,7 +178,7 @@ namespace ReservationAPI.Controllers
             {
                 var role = await _userManager.GetRolesAsync(user);
 
-                IdentityOptions options = new IdentityOptions();
+                IdentityOptions options = new();
                 var tokenDescriptor = new SecurityTokenDescriptor
                 {
                     Subject = new ClaimsIdentity(new Claim[]{
@@ -226,7 +225,7 @@ namespace ReservationAPI.Controllers
                 //get user
                 var socialUser = await _userManager.FindByNameAsync(validation.apiTokenInfo.email);
                 var role = await _userManager.GetRolesAsync(socialUser);
-                IdentityOptions options = new IdentityOptions();
+                IdentityOptions options = new();
 
                 if (socialUser == null)
                 {
@@ -323,66 +322,66 @@ namespace ReservationAPI.Controllers
         }
 
         //GET: /api/User/Profile
-        [HttpGet]
-        [Authorize]
-        [Route("Profile")]
-        public async Task<Object> GetUserProfile()
-        {
-            //auth user -> need to access UserID from claims...
+        //[HttpGet]
+        //[Authorize]
+        //[Route("Profile")]
+        //public async Task<Object> GetUserProfile()
+        //{
+        //    //auth user -> need to access UserID from claims...
 
-            //string userID = User.Claims.FirstOrDefault(c => c.Type == "UserID").Value;
-            //var user = await _userManager.FindByEmailAsync(userID);  //userID je zapravo email u claimsu...
+        //    //string userID = User.Claims.FirstOrDefault(c => c.Type == "UserID").Value;
+        //    //var user = await _userManager.FindByEmailAsync(userID);  //userID je zapravo email u claimsu...
 
-            //var role = await _userManager.GetRolesAsync(user);
+        //    //var role = await _userManager.GetRolesAsync(user);
 
-            //List<UserModel> friends = new List<UserModel>();
-            //UserModel um;
+        //    //List<UserModel> friends = new List<UserModel>();
+        //    //UserModel um;
 
-            //User friend;
+        //    //User friend;
 
-            //foreach (var friendModel in user.Friends)
-            //{
-            //    friend = await _userManager.FindByEmailAsync(friendModel.Email);
+        //    //foreach (var friendModel in user.Friends)
+        //    //{
+        //    //    friend = await _userManager.FindByEmailAsync(friendModel.Email);
 
-            //    um = new UserModel()
-            //    {
-            //        FirstName = friend.FirstName,
-            //        LastName = friend.LastName,
-            //        Email = friend.Email,
-            //        Password = friend.PasswordHash,
-            //        City = friend.City,
-            //        Street = friend.Street,
-            //        Status = role.FirstOrDefault().ToString(),
-            //        PhoneNumber = friend.PhoneNumber,
-            //        Image = friend.Image,
-            //        Friends = new List<UserModel>(),         //prijatelji nece moci da vide prijatelje prijatelja
-            //        Reservations = friend.Reservations,
-            //        PassportNo = friend.PassportNo
-            //    };
+        //    //    um = new UserModel()
+        //    //    {
+        //    //        FirstName = friend.FirstName,
+        //    //        LastName = friend.LastName,
+        //    //        Email = friend.Email,
+        //    //        Password = friend.PasswordHash,
+        //    //        City = friend.City,
+        //    //        Street = friend.Street,
+        //    //        Status = role.FirstOrDefault().ToString(),
+        //    //        PhoneNumber = friend.PhoneNumber,
+        //    //        Image = friend.Image,
+        //    //        Friends = new List<UserModel>(),         //prijatelji nece moci da vide prijatelje prijatelja
+        //    //        Reservations = friend.Reservations,
+        //    //        PassportNo = friend.PassportNo
+        //    //    };
 
-            //    friends.Add(um);
+        //    //    friends.Add(um);
 
-            //}
+        //    //}
 
-            //UserModel returnUser = new UserModel()
-            //{
-            //    FirstName = user.FirstName,
-            //    LastName = user.LastName,
-            //    Email = user.Email,
-            //    Password = user.PasswordHash,
-            //    City = user.City,
-            //    Street = user.Street,
-            //    Status = role.FirstOrDefault().ToString(),
-            //    PhoneNumber = user.PhoneNumber,
-            //    Image = user.Image,
-            //    Friends = friends,
-            //    PassportNo = user.PassportNo
-            //};
+        //    //UserModel returnUser = new UserModel()
+        //    //{
+        //    //    FirstName = user.FirstName,
+        //    //    LastName = user.LastName,
+        //    //    Email = user.Email,
+        //    //    Password = user.PasswordHash,
+        //    //    City = user.City,
+        //    //    Street = user.Street,
+        //    //    Status = role.FirstOrDefault().ToString(),
+        //    //    PhoneNumber = user.PhoneNumber,
+        //    //    Image = user.Image,
+        //    //    Friends = friends,
+        //    //    PassportNo = user.PassportNo
+        //    //};
 
-            //return returnUser;
-            return null;
+        //    //return returnUser;
+        //    return null;
 
-        }
+        //}
 
         #endregion
 
@@ -415,7 +414,7 @@ namespace ReservationAPI.Controllers
             {
                 return await _userService.GetAll();
             }
-            catch (Exception e)
+            catch
             {
                 return new List<UserModel>();
             }
