@@ -29,7 +29,7 @@ namespace ReservationAPI.Extensions
             services.AddScoped<INotification, NotificationService>();
 
             services.AddDbContext<ApplicationDbContext>(
-                options => options.UseSqlServer(config.GetConnectionString("IdentityConnection"))
+                options => options.UseSqlServer(config.GetConnectionString("HomeConnection"))
             );
 
             //file upload
